requirements.txt
pandas~=2.0.2
pytesseract~=0.3.10
pypdfium2~=3.10.0
opencv-python~=4.7.0.72
numpy~=1.24.3
pillow~=9.5.0
utx~=0.0.1
matplotlib~=3.7.1
pypdf2~=3.0.1
pdfrw~=0.4
tqdm~=4.65.0
scipy~=1.10.1
imagesize~=1.4.1
tensorflow~=2.13.0rc1